requirements/base.txt
django==5.0.3
django-environ==0.11.2
django-autoslug==1.9.9
django-cachalot==2.6.2
django-countries==7.5.1
django-taggit==5.0.1
django-cleanup==8.1.0
django-multiselectfield==0.1.12
django-redis==5.4.0
django-cors-headers==4.3.1
django-model-utils==4.4.0
django-celery-email==3.0.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-phonenumber-field==7.3.0
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
pillow==10.2.0
djoser==2.2.2
social-auth-app-django==5.4.0
argon2-cffi==23.1.0
pytz==2024.1
redis==5.0.3
celery==5.3.6
flower==2.0.1
phonenumbers==8.13.32
